CLient/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.17",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
